// frontend/src/generated/model/cooking-appliance-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO containing all information on a cooking appliance.
 */
export interface CookingApplianceDto { 
    /**
     * Unique id of a cooking appliance.
     */
    id: number;
    /**
     * Name of a cooking appliance.
     */
    name: string;
}